rpc: remove logic for generating openapi spec
Since we'll be leveraging gRPC + protobuf instead of a handrolled REST
api we no longer need to maintain the infra for generating an openapi
scpec. In the future we can explore generating an openapi spec from the
proto files themselves as is used in grpc-gateway.
