+++ crates/sui-rpc-api/Cargo.toml
@@ -15,7 +15,6 @@ reqwest.workspace = true
 url.workspace = true
 serde.workspace = true
 serde_json.workspace = true
-serde_yaml.workspace = true
 serde_with.workspace = true
 tap.workspace = true
 thiserror.workspace = true
@@ -25,9 +24,6 @@ itertools.workspace = true
 sui-sdk-types.workspace = true
 sui-sdk-transaction-builder.workspace = true
 prometheus.workspace = true
-openapiv3 = { git = "https://github.com/bmwill/openapiv3.git", rev = "ca4b4845b7c159a39f5c68ad8f7f76cb6f4d6963" }
-schemars.workspace = true
-documented = "0.6.0"
 http.workspace = true
 tower.workspace = true
