+++ crates/sui-rpc-api/src/rest/info.rs
@@ -1,15 +1,13 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::rest::openapi::{ApiEndpoint, OperationBuilder, ResponseBuilder, RouteHandler};
+use super::{ApiEndpoint, RouteHandler};
 use crate::types::NodeInfo;
 use crate::{Result, RpcService};
 use axum::extract::State;
 use axum::Json;
-use documented::Documented;
 
 /// Get basic information about the state of a Node
-#[derive(Documented)]
 pub struct GetNodeInfo;
 
 impl ApiEndpoint<RpcService> for GetNodeInfo {
@@ -21,29 +19,7 @@ impl ApiEndpoint<RpcService> for GetNodeInfo {
         "/"
     }
 
-    fn stable(&self) -> bool {
-        true
-    }
-
-    fn operation(
-        &self,
-        generator: &mut schemars::gen::SchemaGenerator,
-    ) -> openapiv3::v3_1::Operation {
-        OperationBuilder::new()
-            .tag("General")
-            .operation_id("Get NodeInfo")
-            .description(Self::DOCS)
-            .response(
-                200,
-                ResponseBuilder::new()
-                    .json_content::<NodeInfo>(generator)
-                    .build(),
-            )
-            .response(500, ResponseBuilder::new().build())
-            .build()
-    }
-
-    fn handler(&self) -> crate::rest::openapi::RouteHandler<RpcService> {
+    fn handler(&self) -> RouteHandler<RpcService> {
         RouteHandler::new(self.method(), get_node_info)
     }
 }
