+++ crates/sui-rpc-api/src/types.rs
@@ -37,10 +37,9 @@ pub const X_SUI_TIMESTAMP_MS: &str = "x-sui-timestamp-ms";
 
 /// Basic information about the state of a Node
 #[serde_with::serde_as]
-#[derive(Debug, serde::Serialize, serde::Deserialize, schemars::JsonSchema)]
+#[derive(Debug, serde::Serialize, serde::Deserialize)]
 pub struct NodeInfo {
     /// The chain identifier of the chain that this Node is on
-    #[schemars(with = "String")]
     pub chain_id: sui_sdk_types::CheckpointDigest,
 
     /// Human readable name of the chain that this Node is on
@@ -48,28 +47,23 @@ pub struct NodeInfo {
 
     /// Current epoch of the Node based on its highest executed checkpoint
     #[serde_as(as = "sui_types::sui_serde::BigInt<u64>")]
-    #[schemars(with = "crate::rest::_schemars::U64")]
     pub epoch: u64,
 
     /// Checkpoint height of the most recently executed checkpoint
     #[serde_as(as = "sui_types::sui_serde::BigInt<u64>")]
-    #[schemars(with = "crate::rest::_schemars::U64")]
     pub checkpoint_height: u64,
 
     /// Unix timestamp of the most recently executed checkpoint
     #[serde_as(as = "sui_types::sui_serde::BigInt<u64>")]
-    #[schemars(with = "crate::rest::_schemars::U64")]
     pub timestamp_ms: u64,
 
     /// The lowest checkpoint for which checkpoints and transaction data is available
     #[serde_as(as = "Option<sui_types::sui_serde::BigInt<u64>>")]
-    #[schemars(with = "Option<crate::rest::_schemars::U64>")]
     #[serde(skip_serializing_if = "Option::is_none")]
     pub lowest_available_checkpoint: Option<u64>,
 
     /// The lowest checkpoint for which object data is available
     #[serde_as(as = "Option<sui_types::sui_serde::BigInt<u64>>")]
-    #[schemars(with = "Option<crate::rest::_schemars::U64>")]
     #[serde(skip_serializing_if = "Option::is_none")]
     pub lowest_available_checkpoint_objects: Option<u64>,
     pub software_version: std::borrow::Cow<'static, str>,
@@ -141,7 +135,7 @@ pub struct CheckpointResponse {
     pub contents_bcs: Option<Vec<u8>>,
 }
 
-#[derive(Clone, Debug, Default, serde::Serialize, serde::Deserialize, schemars::JsonSchema)]
+#[derive(Clone, Debug, Default, serde::Serialize, serde::Deserialize)]
 pub struct GetCheckpointOptions {
     /// Request `CheckpointSummary` be included in the response
     ///
@@ -242,7 +236,7 @@ pub struct TransactionResponse {
     pub timestamp_ms: Option<u64>,
 }
 
-#[derive(Clone, Debug, Default, serde::Serialize, serde::Deserialize, schemars::JsonSchema)]
+#[derive(Clone, Debug, Default, serde::Serialize, serde::Deserialize)]
 pub struct GetTransactionOptions {
     /// Request `Transaction` be included in the response
     ///
@@ -328,7 +322,7 @@ impl GetTransactionOptions {
 }
 
 /// Options for the execute transaction endpoint
-#[derive(Clone, Debug, Default, serde::Serialize, serde::Deserialize, schemars::JsonSchema)]
+#[derive(Clone, Debug, Default, serde::Serialize, serde::Deserialize)]
 pub struct ExecuteTransactionOptions {
     /// Request `TransactionEffects` be included in the Response.
     ///
@@ -432,7 +426,7 @@ pub enum EffectsFinality {
     QuorumExecuted,
 }
 
-#[derive(Clone, Debug, Default, serde::Serialize, serde::Deserialize, schemars::JsonSchema)]
+#[derive(Clone, Debug, Default, serde::Serialize, serde::Deserialize)]
 pub struct GetFullCheckpointOptions {
     /// Request `CheckpointSummary` be included in the response
     ///
